requirements.txt
streamlit
langchain
langchain_groq
lanchain_community
langchain_core
python-dotenv
langchain_huggingface